amidala_company/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
